+++ src/app/(route)/mypage/layout.tsx
@@ -1,6 +1,6 @@
 import type { PropsWithChildren } from 'react';
 
-import { FooterLayout } from '@/components/layout/footer';
+import { FooterLayout } from '@/components/layout/Footer';
 
 export default function MyPageLayout({ children }: PropsWithChildren) {
   return <FooterLayout>{children}</FooterLayout>;
