+++ src/app/(route)/report/layout.tsx
@@ -1,7 +1,7 @@
 import type { PropsWithChildren } from 'react';
 
-import { FooterLayout } from '@/components/layout/footer';
-import { TopBar } from '@/components/layout/topBar';
+import { FooterLayout } from '@/components/layout/Footer';
+import { TopBar } from '@/components/layout/TopBar';
 
 export default function ReportLayout({ children }: PropsWithChildren) {
   return (
